+++ src/index.ts
@@ -1,7 +1,6 @@
 export * from './main/spotifyApi';
 
 export type * from './apis/index.types';
-export type * from './helpers/index.types';
 export type * from './main/spotifyAxios';
 export type * from './openapi';
 export type * from './types';
